Blind - Auction
One of the main perks of blind bidding from a seller's perspective is that it can lead to a much higher sale price. Since buyers cannot see what anyone else has bid, they might end up offering a price well over asking in an effort to outbid the competition and secure the home.