Guess-the-number-final
A number guessing game is a simple guessing game where a user is supposed to guess a number between 0 and N in a maximum of 10 attempts  and in hard level maximum 5 attempts. The game will end after 10 attempts and if the player failed to guess the number, and then he loses the game.